% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{anneal_cv_anneal}
\alias{anneal_cv_anneal}
\title{Generate (# of seasons)-fold CV data.}
\usage{
anneal_cv_anneal(
  data,
  .datetime,
  .observation,
  .smoothed_obs,
  tmax_offset = 0,
  season_len = 365,
  outer_digest_args = list(),
  inner_digest_args = list(),
  anneal_args = list()
)
}
\arguments{
\item{data}{tsibble. Your data, with a column of smoothed observations.}

\item{.datetime}{\code{data}'s column of datetimes, i.e. your tsibble's \code{index}.}

\item{.observation}{\code{data}'s column of original observations.}

\item{.smoothed_obs}{\code{data}'s column of smoothed observations.}

\item{tmax_offset}{int. The offset from your latest observation to consider
as "tmax".}

\item{season_len}{int. The season length.}

\item{outer_digest_args}{list. Any arguments to override the defaults for
the outer digest loop.}

\item{inner_digest_args}{list. Any arguments to override the defaults for
the inner digest loop.}

\item{anneal_args}{list. Any arguments to override the defaults for \code{anneal}.}
}
\value{
list. All of the arguments used from annealing, i.e.:
- .smoothed_obs_name,
- tmax_offset,
- (outer/inner/anneal)-args
Also, annealing information. A list, with an element for each outer-
loop digest fragment.
- anneal_output,
- test_data_outer_fragment (i.e. the outer-loop digest fragment),
- k_test, i.e. the k of the test_data_outer_fragment
}
\description{
Generate CV data using a particular set of hyperparameters.
}
\details{
This function may be plugged into a grid search for hyperparameters.
The \code{.datetime} and \code{.observation} columns should stay the
same across calls. Most of the other args could reasonably be CV'ed. A
few things worth mentioning:

\code{tmax_offset} is useful for seeing how a specific method / set of hyperparams
performs at different times of the season. For instance, for daily data, you
could use values such as -14, -7, 0, 7, 14, to help evaluate +- two weeks.

The \code{loss_fn} arg for \code{anneal}, which you would pass via \code{anneal_args},
needs to be a list (e.g. list(myfun = myfun)) for reasons.

See the source code for the default args in the o/i/a-args lists.

The rationale behind this function is that the smoothed data was locally
smoothed, so CV should just work okay on a seasonal basis -- Jan 13 of
this year has no bearing on Jan 13 of next year, w.r.t smoothing. Ergo,
there is a two-step digest process:
\enumerate{
\item An outer, broad-range digest, yielding a set of k (#-seasons of) "data"s.
\item An inner, more restricted digest, yielding a set of k (#-seasons of)
"fragments".
On each interation i, the outer "data" fragment k == i is used, and the
inner fragments k != i are used.
Each outer fragment is padded back to t_0 with NAs for compatibility with the
current implementation of \code{anneal}.
}

I've padded the data with another season of NAs. That means that the k == 1
fragment is the current season, while in a typical \code{digest}, k == 1 is prior
season (because the current season isn't even included). This should yield
a bit more power when \code{tmax_offset} < 0 and/or the annealed k == 1 fragment
has a positive offset, and I couldn't think of a downside. I may make it
optional, though.
}
