% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{calc_ortho_vec}
\alias{calc_ortho_vec}
\title{Calculate a vector orthogonal to the normalized average difference of your time series.}
\usage{
calc_ortho_vec(ts, n, scaler = 1, wts = NULL)
}
\arguments{
\item{ts}{The time series. Should be smooth/smoothed. e.g. dat$readings}

\item{n}{The number of values at the end of the value column to include. Constraint: 1 < n < len(value)}

\item{scaler}{Scale your y's so that they're on equal footing w/ dx.}

\item{wts}{Optional weights for each difference.}
}
\value{
tibble (1x2) Normalized orthogonal vector.
}
\description{
The steps are:
\enumerate{
\item Take the last n values of your ts.
\item Take their difference.
\item Perform a 90 degree counter-clockwise rotation on (diff, 1) vectors.
\item Normalize each rotated vector.
\item Optionally weight each vector.
\item Add these vectors.
\item Normalize the total vector.
}
}
