% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{digest}
\alias{digest}
\title{Fragment your time series by seasons, with an overlap.}
\usage{
digest(
  data,
  n_overlap,
  season_len,
  n_future_steps = 120,
  include_partial_overlap = TRUE
)
}
\arguments{
\item{data}{The data.}

\item{n_overlap}{The number of time points to overlap. Must be positive integer.}

\item{season_len}{The number of time points per season. Must be positive integer.}

\item{n_future_steps}{The number of time points for fragments to extend beyond your latest observation.}

\item{include_partial_overlap}{Whether to include the final fragment if there is only partial overlap.}
}
\value{
A tibble of fragments containing cols:
idx:     The index w.r.t the original data.
k:       The fragment (1 = 1 season back, etc.).
adj_idx: The fragment, aligned to the latest data.
}
\description{
A "fragment" is 1 or more prior seasons' of data, plus an overlap for annealing.
This function is of the same ilk as \code{stretch_tsibble} from the \code{tsibble} package.
}
\details{
Note that potentially "empty" fragments may be returned (e.g. no data
collected for a given year). To remove these, you can pass the output to
\code{trim_fragments_na}.
}
