% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{mark_long_na_sequences}
\alias{mark_long_na_sequences}
\title{Create a boolean vector indicating long sequences of NAs.}
\usage{
mark_long_na_sequences(observations, maxgap)
}
\arguments{
\item{observations}{The observations to generate the boolean vector for.}

\item{maxgap}{The longest sequence of NAs to lump in (boolean-ly) with the non-NAs.}
}
\value{
A boolean vector of length(observations).
TRUE  = The observation (well, lack thereof) is part of a sequence of NAs
that exceeds maxgap in length.
FALSE = Either observations, or sequences of NAs <= maxgap.
}
\description{
This function is useful for subsequent use in filtering out these long
sequences from your data.
}
\details{
Note that the long sequences of NAs are the "TRUE"s, so to filter,
be mindful of using the logical not \code{!}.
}
