% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_anneal_fragment}
\alias{plot_anneal_fragment}
\title{Plot an annealed fragment against the smoothed data.}
\usage{
plot_anneal_fragment(
  data,
  x,
  y,
  anneal_output,
  fragment_obs_col_name,
  k = 1,
  truncate = TRUE
)
}
\arguments{
\item{data}{.}

\item{x}{The datetime column.}

\item{y}{The column of smoothed (fitted) observations.}

\item{anneal_output}{Output of \code{anneal}.}

\item{fragment_obs_col_name}{The (possibly smoothed) fragments' observations column name.}

\item{k}{Which fragment number to plot.}

\item{truncate}{Truncate the plot to \code{data}'s most recent observation?}
}
\description{
This plot provides a way to assess how good the annealed fragment fits
the data.
}
\details{
Truncating is useful for excluding the parts of fragments that extend beyond
your most recent observation, and therefore do not factor into the
loss function.
}
