% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{plot_anneal_losses}
\alias{plot_anneal_losses}
\title{Plot the losses output of anneal.}
\usage{
plot_anneal_losses(anneal_out)
}
\arguments{
\item{anneal_out}{Output of \code{anneal}.}
}
\description{
Plot the losses output of anneal.
}
