% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_anneal_original_fragments}
\alias{plot_anneal_original_fragments}
\title{Plot the unshifted fragments against the original data.}
\usage{
plot_anneal_original_fragments(
  data,
  x,
  y,
  anneal_output,
  fragment_obs_col_name,
  t_max = NULL
)
}
\arguments{
\item{data}{The data.}

\item{x}{The datetime column.}

\item{y}{The (possibly smoothed) observations.}

\item{anneal_output}{Output of \code{anneal}.}

\item{fragment_obs_col_name}{The (possibly smoothed) fragments' observations column name.}

\item{t_max}{The upper xlim.}
}
\description{
Useful for seeing what a naive seasonal forecast would look like,
and comparing this to what \code{anneal} gives you via
\code{plot_min_loss_fragments}.
}
\details{
Setting t_max at, e.g. \code{lubridate::today() + 30}, is useful for
visual comparisons with \code{plot_min_loss_fragments} (with the same
t_max).
}
