% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{plot_data_with_ortho_vec}
\alias{plot_data_with_ortho_vec}
\title{Plot a smoothed time series with the orthogonal vector to its head overlaid.}
\usage{
plot_data_with_ortho_vec(
  data,
  index,
  value,
  n,
  normed_total_vec,
  arrow_scale = 1
)
}
\arguments{
\item{data}{Your data containing a smoothed time series.}

\item{index}{The date/time column.}

\item{value}{The column of smoothed values.}

\item{n}{The \code{n} used in \code{calc_ortho_vec}.}

\item{normed_total_vec}{Output of \code{calc_ortho_vec}.}

\item{arrow_scale}{Scale of arrow. \code{1} = 1 unit of index, 1 unit of value.}
}
\description{
The orthogonal vector comes from \code{calc_ortho_vec}.
The smoothed time series is, e.g. a \code{lowess} output.
}
