% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{plot_min_loss_fragments}
\alias{plot_min_loss_fragments}
\title{Plot the min-loss shifted fragments against the original data.}
\usage{
plot_min_loss_fragments(data, x, y, anneal_output, t_max = NULL)
}
\arguments{
\item{data}{The data.}

\item{x}{The datetime column.}

\item{y}{The (possibly smoothed) observations.}

\item{anneal_output}{Output of \code{anneal}.}

\item{t_max}{The upper xlim.}
}
\description{
Useful for seeing what a forecast would look like,
and comparing this to using the naive seasonal via
\code{plot_original_fragments}.
}
\details{
Setting t_max at, e.g. \code{lubridate::today() + 30}, is useful for
visual comparisons with \code{plot_min_loss_fragments} (with the same
t_max).
}
