% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ortho_vec}
\alias{plot_ortho_vec}
\title{Plot the output of \code{calc_ortho_vec}.}
\usage{
plot_ortho_vec(normed_total_vec)
}
\arguments{
\item{normed_total_vec}{Output of \code{calc_ortho_vec}.}
}
\description{
Plot the output of \code{calc_ortho_vec}.
}
