% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_fns.R
\name{recency_weighted_rmse}
\alias{recency_weighted_rmse}
\title{Calculate the root mean squared error of two sequences, weighted by recency.}
\usage{
recency_weighted_rmse(a, b)
}
\arguments{
\item{a}{Sequence 1}

\item{b}{Sequence 2}
}
\value{
Recency-weighted RMSE
}
\description{
Weights are based on the length(a) linear sequence over \link{0,1}.
At time t, the formulation is: ((a_t - b_t) ^ 2) * wt_t.
}
\details{
Intended to be passed as the \code{loss_fn} arg of \code{anneal}.

In the context of this package, this punishes strong deviations near the
head of your time series. For instance, a sudden flattening in the last
few observations may be important.

This is not always the best option, however, and results should be
inspected.
}
