% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{rmse}
\alias{rmse}
\title{Calculate the root mean squared error of two sequences.}
\usage{
rmse(a, b)
}
\arguments{
\item{a}{Sequence 1}

\item{b}{Sequence 2}
}
\value{
RMSE
}
\description{
Calculate the root mean squared error of two sequences.
}
