% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_fns.R
\name{slope_weighted_rmse}
\alias{slope_weighted_rmse}
\title{Calculate the root mean squared error of two sequences, weighted by slope.}
\usage{
slope_weighted_rmse(a, b)
}
\arguments{
\item{a}{Sequence 1}

\item{b}{Sequence 2}
}
\value{
Slope-weighted RMSE
}
\description{
Weights are based on sequence \code{b}'s \code{difference}. At time t,
the formulation is: ((a_t - b_t) * (b_t - b_{t-1})) ^ 2.
}
\details{
Intended to be passed as the \code{loss_fn} arg of \code{anneal}.

In the context of this package, \code{b} makes more sense as your fragment,
because the "current" data, \code{a}, may have a misleading trajectory for
its smoothed fit (e.g. an imminent upturn / other inflection point
isn't reflected because it hasn't happened yet.).

This weighting means that the flat parts / inflection points of
the series have minimal influence. Ergo, the sequences' alignment
should be driven by the spring/fall temperature changes, not the
extremeness of summer/winter.

This is not always the best option, however, and results should be
inspected.
}
