% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{trim_fragments_na}
\alias{trim_fragments_na}
\title{Trim digest fragments of leading/trailing NAs}
\usage{
trim_fragments_na(digest, .col, left = TRUE, right = TRUE)
}
\arguments{
\item{digest}{Output of \code{digest}.}

\item{.col}{The column to search through for NAs.}

\item{left}{Trim left side?}

\item{right}{Trim right side?}
}
\value{
digest with trimmed fragments.
}
\description{
This is useful for cases where one/both ends of your fragment(s)'s
original observations contain NA's that you don't want to have \code{predict}ed with
your \code{loess_fit} in \code{anneal}, such as a large gap beyond the end(s) of the
fragment until the next non-NA observation (i.e. you don't trust the fit
for these NA's).
}
\details{
This function always removes fragments containing only NA observations.
As such, setting both \code{left} and \code{right} to \code{FALSE} will remove
only these all-NA fragments.
}
