% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{anneal}
\alias{anneal}
\title{Assumptions:
\enumerate{
\item Object passed to \code{loess_fit} had the formula \verb{[var1] ~ [var2]}, e.g. \code{obs ~ date_time}.
\item Object passed to \code{data} includes the \verb{[var1]} col from Assumption #1 with the same name, e.g. \code{obs}.
}}
\usage{
anneal(
  data,
  fitted_obs_col_name,
  digest,
  resolution,
  range_start,
  range_end,
  ortho_vec,
  loess_fit,
  loss_fn
)
}
\arguments{
\item{data}{A tsibble. Cannot have multiple elements in key, if key exists.}

\item{fitted_obs_col_name}{The column name in \code{data} of fitted observations from \verb{predict(loess_fit, new_data = <your data idx>)}.}

\item{digest}{Output of \code{anneal::digest} for smoothed data.}

\item{resolution}{The step size (along x-axis) during search. Currently must equally divide 1 (e.g. 1, 0.2, or 0.5).}

\item{range_start}{Start of range to search over.}

\item{range_end}{End of range to search over.}

\item{ortho_vec}{Output of \code{anneal_calc_ortho_vec} for smoothed data.}

\item{loess_fit}{Fitted output of \code{loess} on your data.}

\item{loss_fn}{f(seq, seq) The loss function.}
}
\value{
list(
shifted_data = tibble,
losses       = tibble)
}
\description{
Assumptions:
\enumerate{
\item Object passed to \code{loess_fit} had the formula \verb{[var1] ~ [var2]}, e.g. \code{obs ~ date_time}.
\item Object passed to \code{data} includes the \verb{[var1]} col from Assumption #1 with the same name, e.g. \code{obs}.
}
}
