% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anneal.R
\name{digest}
\alias{digest}
\title{Fragment your time series by seasons, with an overlap.}
\usage{
digest(
  data,
  .datetime,
  .observation,
  n_overlap,
  season_len,
  n_future_steps = 120,
  max_na_sequence = Inf,
  min_fragment_len = 2
)
}
\arguments{
\item{data}{tsibble. The data. Must not contain gaps.}

\item{.datetime}{Your datetime column. The tsibble's "index".}

\item{.observation}{Your observations column.}

\item{n_overlap}{The number of time points to overlap. Must be positive integer.}

\item{season_len}{The number of time points per season. Must be positive integer.}

\item{n_future_steps}{The number of time points for fragments to extend beyond your latest observation.}

\item{max_na_sequence}{The largest series of NAs (i.e. gap between observations) before filtering out.}

\item{min_fragment_len}{The minimum length of a fragment to include in output.}
}
\value{
A tibble of fragments augmented with cols:
idx:          The index w.r.t the original data.
k:            The fragment (1 = 1 season back, etc.).
adj_idx:      The fragment's index aligned to the latest data.
adj_datetime: The fragment's datetime aligned to the latest data.
}
\description{
A "fragment" is 1 or more prior seasons' of data. From a baseline
of k seasons before your latest observation:
\enumerate{
\item It includes a "future" portion of \code{n_future_steps} observations,
for forecasting purposes.
\item It also includes a "past" portion of \code{n_overlap} observations,
for annealing purposes.
This function is of the same ilk as \code{stretch_tsibble} from the \code{tsibble} package.
}
}
\details{
The data cannot contain gaps, because this function relies on indexing.
You can be sure that it doesn't by making it a \code{tsibble} and checking
for gaps.

Your \code{n_future_steps} should be longer than your desired forecast horizon,
because \code{anneal} will potentially shift it backwards. A sensible lower bound
is (your desired forecast horizon) + (your max backwards shift in \code{anneal}).

\code{max_na_sequence} is used to remove long sequences of missing data.
Because \code{predict} can predict anywhere (i.e. the observation can be NA),
this prevents them from being \code{predict}ed in \code{anneal} and thereby
prevents them from factoring into \code{anneal}'s loss calculation.

Fragments shorter than \code{min_fragment_len}, after any trimming via
\code{max_na_sequence} (though the fragment may still contain or
have at its ends or even be entirely NA observations,
which count towards the length) are removed.
It defaults to 2, because:
\itemize{
\item What can you do with a fragment of length 1?
\item tsibble doesn't like it.
}
}
